Configure logging level.

This will configure the logger level. This allows to turn DEBUG, INFO, WARN, ERROR, and SEVERE messages on and off.

Usage

logger.setLevel(level)

Arguments

level

the level of the message. One of "ALL", "DEBUG", "INFO", "WARN", "ERROR", "SEVERE", or "OFF".

Value

When logger level is set, the previous level is returned invisibly. This can be passed to logger.setLevel() to restore the previous level.

Details

Note that this controls printing of messages and does not change other behavior. In particular, suppressing SEVERE by setting the level to "OFF" does not prevent logger.severe() from signaling an error (and terminating the program if logger.setQuitOnSevere(TRUE)).

Author

Rob Kooper

Examples

if (FALSE) { # \dontrun{
logger.setLevel('DEBUG')

# Temporarily turn logger off
old_logger_level <- logger.setLevel("OFF")
  # code here
logger.setLevel(old_logger_level)
} # }