Configure logging output filename.

The name of the file where the logging information should be written to.

Usage

logger.setOutputFile(filename)

Arguments

filename

the file to send the log messages to (or NA to not write to file)

Value

Invisibly, the previously set filename. This can be used to restore settings after a temporary change.

Author

Rob Kooper

Examples

if (FALSE) { # \dontrun{
logger.setOutputFile('pecan.log')
} # }